+++ src/components/Home.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import ShopButton from './ShopButton';
-import CategoriesList from './CategoriesList';
+
 import SearchBar from './SearchBar';
 
 class Home extends Component {
@@ -9,13 +9,7 @@ class Home extends Component {
     return (
       <section>
         <SearchBar />
-        <CategoriesList />
         <ShopButton />
-        <ul>
-          <li data-testid="home-initial-message">
-            Digite algum termo de pesquisa ou escolha uma categoria.
-          </li>
-        </ul>
       </section>
     );
   }
