+++ src/components/SearchBar.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import * as api from '../services/api';
 import ProductCard from './ProductCard';
+import Categories from './CategoriesList';
 
 class SearchBar extends Component {
   constructor() {
@@ -9,6 +10,7 @@ class SearchBar extends Component {
       product: [],
       query: '',
     };
+    this.handleClick = this.handleClick.bind(this);
     this.handleSearch = this.handleSearch.bind(this);
     this.handleChange = this.handleChange.bind(this);
   }
@@ -23,6 +25,13 @@ class SearchBar extends Component {
     const { name, value } = target;
     this.setState({
       [name]: value,
+    });
+  }
+
+  async handleClick({ target }) {
+    const data = await api.getProductsFromCategoryAndQuery(target.key, target.value);
+    this.setState({
+      product: data.results,
     });
   }
 
@@ -47,6 +56,10 @@ class SearchBar extends Component {
         >
           Button
         </button>
+        <Categories onClick={ this.handleClick } />
+        <h3 data-testid="home-initial-message">
+          Digite algum termo de pesquisa ou escolha uma categoria.
+        </h3>
         { product.map((item) => <ProductCard key={ item.id } product={ item } />) }
       </section>
     );
